Major changes to the integrator interface.
One acceleration variable is used per property that has to be stepped.
This will reduce the number of unnecessary arrays created for the
particle array.

We loose the ability to inspect each calc's output separately.The benefits of
accumulated arrays are a much simpler integrator interface and standard acceleration
variable names.