Chanigng the output returned from the solver util's load function.
Currently, the output is a dictionary with two keys:

(1) solver_data : All solver related data like time, time step and iteration count.
(2) arrays : A dict of keyed on particle array name and the particle array as value.

Changes to the viewer:
(1) Fixing a bug that would load the arrays of any running solver instead of the saved files.
(2) Use the solver.util's load function to read the saved data.