Adding the default parallel manager thanks to Prabhu Ramachandran.
Currently, a trivial parallel manager is implemented which shares all particle
data with all processors taking care to mark them as remote.

API's that have changed:

(a) base.Particles has a parallel_manager instance that is None by default.
    The application sets the parallel manager for the particles at the time of
    particle creation.

(b) base.CellManager now calls a rebin on the particles instead of a `cells_update`
    This works in serial and in parallel where we need to insert particles that have
    either moved into out region or are remote for neighbor purposes. In serial runs,
    the jump tolerance problem is eliminated as it is never checked.

(c) Application accepts a new command line argument "--parallel-mode" which is used to
    switch between the available parallel managers. Currently, only the simple parallel
    manager is implemented. Activate this explicitly using --parallel-mode=simple
    This is also the default.

Tests:
Updating the script to run serial and parallel runs and compare the results.

Issues:

(a) The space filling curves based distribution function results in an error.
    Currently, for the simple manager, we use the "auto" method for distributing
    particles.