BUG: Forcing cell manager to rebin all particles for variable h
The cell sizes are computed at each update before re-binning all particles.

The specification of the minimum and maximum cell sizes works logically now.
The default cell size is scale_fac*max_h and if this lies outside the
range [min_cell_size, max_cell_size], it is clamped to the nearest value.