BUG: Adding an operation to reset 'h' for the ADKE procedure.
The ADKE procedure estimates the pilot density based on a fixed
smoothing length 'h0'. We need to reset this smoothing length before
estimating the pilot density. Currently, the old smoothing length was being used
to find the effective radius which led to oscillations in the smoothing length profile
in regions of sharp density variations.